+++ packages/transport/src/events/MessageWasReadAtChangedEvent.ts
@@ -1,4 +1,4 @@
-import { Message } from "../modules/messages/local/Message";
+import { Message } from "../modules";
 import { TransportDataEvent } from "./TransportDataEvent";
 
 export class MessageWasReadAtChangedEvent extends TransportDataEvent<Message> {
