+++ packages/transport/src/events/PeerRelationshipTemplateLoadedEvent.ts
@@ -1,4 +1,4 @@
-import { RelationshipTemplate } from "../modules/relationshipTemplates/local/RelationshipTemplate";
+import { RelationshipTemplate } from "../modules";
 import { TransportDataEvent } from "./TransportDataEvent";
 
 export class PeerRelationshipTemplateLoadedEvent extends TransportDataEvent<RelationshipTemplate> {
